requirements.txt
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.3
boto3==1.28.73
botocore==1.31.73
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
Flask==3.0.0
Flask-SocketIO==5.3.6
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
s3transfer==0.7.0
simple-websocket==1.0.0
six==1.16.0
urllib3==2.0.7
websocket-client==1.6.4
Werkzeug==3.0.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
